+++ lib/pleroma/web/mastodon_api/views/status_view.ex
@@ -294,7 +294,8 @@ defmodule Pleroma.Web.MastodonAPI.StatusView do
         expires_at: expires_at,
         direct_conversation_id: direct_conversation_id,
         thread_muted: thread_muted?,
-        emoji_reactions: emoji_reactions
+        emoji_reactions: emoji_reactions,
+        local_only: Activity.local_only?(activity)
       }
     }
   end