+++ test/web/mastodon_api/views/status_view_test.exs
@@ -198,7 +198,8 @@ defmodule Pleroma.Web.MastodonAPI.StatusViewTest do
         expires_at: nil,
         direct_conversation_id: nil,
         thread_muted: false,
-        emoji_reactions: []
+        emoji_reactions: [],
+        local_only: false
       }
     }